+++ sdk/kubernetesconfiguration/azure-mgmt-kubernetesconfiguration/azure/mgmt/kubernetesconfiguration/v2020_07_01_preview/models/_models_py3.py
@@ -1,5 +1,5 @@
+# pylint: disable=line-too-long,useless-suppression
 # coding=utf-8
-# pylint: disable=too-many-lines
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
@@ -10,10 +10,9 @@
 import datetime
 from typing import Any, Dict, List, Optional, TYPE_CHECKING, Union
 
-from ... import _serialization
+from .._utils import serialization as _serialization
 
 if TYPE_CHECKING:
-    # pylint: disable=unused-import,ungrouped-imports
     from .. import models as _models
 
 
@@ -66,7 +65,7 @@ class ComplianceStatus(_serialization.Model):
          ~azure.mgmt.kubernetesconfiguration.v2020_07_01_preview.models.MessageLevelType
         """
         super().__init__(**kwargs)
-        self.compliance_state = None
+        self.compliance_state: Optional[Union[str, "_models.ComplianceStateType"]] = None
         self.last_config_applied = last_config_applied
         self.message = message
         self.message_level = message_level
@@ -110,15 +109,15 @@ class ConfigurationIdentity(_serialization.Model):
          ~azure.mgmt.kubernetesconfiguration.v2020_07_01_preview.models.ResourceIdentityType
         """
         super().__init__(**kwargs)
-        self.principal_id = None
-        self.tenant_id = None
+        self.principal_id: Optional[str] = None
+        self.tenant_id: Optional[str] = None
         self.type = type
 
 
 class ErrorDefinition(_serialization.Model):
     """Error definition.
 
-    All required parameters must be populated in order to send to Azure.
+    All required parameters must be populated in order to send to server.
 
     :ivar code: Service specific error code which serves as the substatus for the HTTP error code.
      Required.
@@ -170,7 +169,7 @@ class ErrorResponse(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.error = None
+        self.error: Optional["_models.ErrorDefinition"] = None
 
 
 class Resource(_serialization.Model):
@@ -210,9 +209,9 @@ class Resource(_serialization.Model):
          ~azure.mgmt.kubernetesconfiguration.v2020_07_01_preview.models.SystemData
         """
         super().__init__(**kwargs)
-        self.id = None
-        self.name = None
-        self.type = None
+        self.id: Optional[str] = None
+        self.name: Optional[str] = None
+        self.type: Optional[str] = None
         self.system_data = system_data
 
 
@@ -232,30 +231,8 @@ class ProxyResource(Resource):
     :vartype system_data: ~azure.mgmt.kubernetesconfiguration.v2020_07_01_preview.models.SystemData
     """
 
-    _validation = {
-        "id": {"readonly": True},
-        "name": {"readonly": True},
-        "type": {"readonly": True},
-    }
 
-    _attribute_map = {
-        "id": {"key": "id", "type": "str"},
-        "name": {"key": "name", "type": "str"},
-        "type": {"key": "type", "type": "str"},
-        "system_data": {"key": "systemData", "type": "SystemData"},
-    }
-
-    def __init__(self, *, system_data: Optional["_models.SystemData"] = None, **kwargs: Any) -> None:
-        """
-        :keyword system_data: Top level metadata
-         https://github.com/Azure/azure-resource-manager-rpc/blob/master/v1.0/common-api-contracts.md#system-metadata-for-all-azure-resources.
-        :paramtype system_data:
-         ~azure.mgmt.kubernetesconfiguration.v2020_07_01_preview.models.SystemData
-        """
-        super().__init__(system_data=system_data, **kwargs)
-
-
-class ExtensionInstance(ProxyResource):  # pylint: disable=too-many-instance-attributes
+class ExtensionInstance(ProxyResource):
     """The Extension Instance object.
 
     Variables are only populated by the server, and will be ignored when sending a request.
@@ -402,12 +379,12 @@ class ExtensionInstance(ProxyResource):  # pylint: disable=too-many-instance-att
         self.scope = scope
         self.configuration_settings = configuration_settings
         self.configuration_protected_settings = configuration_protected_settings
-        self.install_state = None
+        self.install_state: Optional[Union[str, "_models.InstallStateType"]] = None
         self.statuses = statuses
-        self.creation_time = None
-        self.last_modified_time = None
-        self.last_status_time = None
-        self.error_info = None
+        self.creation_time: Optional[str] = None
+        self.last_modified_time: Optional[str] = None
+        self.last_status_time: Optional[str] = None
+        self.error_info: Optional["_models.ErrorDefinition"] = None
         self.identity = identity
 
 
@@ -437,8 +414,8 @@ class ExtensionInstancesList(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.value = None
-        self.next_link = None
+        self.value: Optional[List["_models.ExtensionInstance"]] = None
+        self.next_link: Optional[str] = None
 
 
 class ExtensionInstanceUpdate(_serialization.Model):
@@ -609,7 +586,7 @@ class ResourceProviderOperation(_serialization.Model):
         super().__init__(**kwargs)
         self.name = name
         self.display = display
-        self.is_data_action = None
+        self.is_data_action: Optional[bool] = None
 
 
 class ResourceProviderOperationDisplay(_serialization.Model):
@@ -687,7 +664,7 @@ class ResourceProviderOperationList(_serialization.Model):
         """
         super().__init__(**kwargs)
         self.value = value
-        self.next_link = None
+        self.next_link: Optional[str] = None
 
 
 class Result(_serialization.Model):
@@ -788,7 +765,7 @@ class ScopeNamespace(_serialization.Model):
         self.target_namespace = target_namespace
 
 
-class SourceControlConfiguration(ProxyResource):  # pylint: disable=too-many-instance-attributes
+class SourceControlConfiguration(ProxyResource):
     """The SourceControl Configuration object returned in Get & Put response.
 
     Variables are only populated by the server, and will be ignored when sending a request.
@@ -929,12 +906,12 @@ class SourceControlConfiguration(ProxyResource):  # pylint: disable=too-many-ins
         self.operator_params = operator_params
         self.configuration_protected_settings = configuration_protected_settings
         self.operator_scope = operator_scope
-        self.repository_public_key = None
+        self.repository_public_key: Optional[str] = None
         self.ssh_known_hosts_contents = ssh_known_hosts_contents
         self.enable_helm_operator = enable_helm_operator
         self.helm_operator_properties = helm_operator_properties
-        self.provisioning_state = None
-        self.compliance_status = None
+        self.provisioning_state: Optional[Union[str, "_models.ProvisioningStateType"]] = None
+        self.compliance_status: Optional["_models.ComplianceStatus"] = None
 
 
 class SourceControlConfigurationList(_serialization.Model):
@@ -963,8 +940,8 @@ class SourceControlConfigurationList(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.value = None
-        self.next_link = None
+        self.value: Optional[List["_models.SourceControlConfiguration"]] = None
+        self.next_link: Optional[str] = None
 
 
 class SystemData(_serialization.Model):
@@ -1010,9 +987,9 @@ class SystemData(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.created_by = None
-        self.created_by_type = None
-        self.created_at = None
-        self.last_modified_by = None
-        self.last_modified_by_type = None
-        self.last_modified_at = None
+        self.created_by: Optional[str] = None
+        self.created_by_type: Optional[str] = None
+        self.created_at: Optional[datetime.datetime] = None
+        self.last_modified_by: Optional[str] = None
+        self.last_modified_by_type: Optional[str] = None
+        self.last_modified_at: Optional[datetime.datetime] = None