+++ sdk/kubernetesconfiguration/azure-mgmt-kubernetesconfiguration/azure/mgmt/kubernetesconfiguration/v2021_11_01_preview/models/_models_py3.py
@@ -1,5 +1,5 @@
+# pylint: disable=line-too-long,useless-suppression,too-many-lines
 # coding=utf-8
-# pylint: disable=too-many-lines
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
@@ -8,18 +8,11 @@
 # --------------------------------------------------------------------------
 
 import datetime
-import sys
-from typing import Any, Dict, List, Optional, TYPE_CHECKING, Union
+from typing import Any, Dict, List, Literal, Optional, TYPE_CHECKING, Union
 
-from ... import _serialization
-
-if sys.version_info >= (3, 8):
-    from typing import Literal  # pylint: disable=no-name-in-module, ungrouped-imports
-else:
-    from typing_extensions import Literal  # type: ignore  # pylint: disable=ungrouped-imports
+from .._utils import serialization as _serialization
 
 if TYPE_CHECKING:
-    # pylint: disable=unused-import,ungrouped-imports
     from .. import models as _models
 
 
@@ -53,9 +46,9 @@ class Resource(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.id = None
-        self.name = None
-        self.type = None
+        self.id: Optional[str] = None
+        self.name: Optional[str] = None
+        self.type: Optional[str] = None
 
 
 class ProxyResource(Resource):
@@ -73,22 +66,6 @@ class ProxyResource(Resource):
      "Microsoft.Storage/storageAccounts".
     :vartype type: str
     """
-
-    _validation = {
-        "id": {"readonly": True},
-        "name": {"readonly": True},
-        "type": {"readonly": True},
-    }
-
-    _attribute_map = {
-        "id": {"key": "id", "type": "str"},
-        "name": {"key": "name", "type": "str"},
-        "type": {"key": "type", "type": "str"},
-    }
-
-    def __init__(self, **kwargs: Any) -> None:
-        """ """
-        super().__init__(**kwargs)
 
 
 class ClusterScopeSettings(ProxyResource):
@@ -192,7 +169,7 @@ class ComplianceStatus(_serialization.Model):
          ~azure.mgmt.kubernetesconfiguration.v2021_11_01_preview.models.MessageLevelType
         """
         super().__init__(**kwargs)
-        self.compliance_state = None
+        self.compliance_state: Optional[Union[str, "_models.ComplianceStateType"]] = None
         self.last_config_applied = last_config_applied
         self.message = message
         self.message_level = message_level
@@ -243,8 +220,8 @@ class ErrorAdditionalInfo(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.type = None
-        self.info = None
+        self.type: Optional[str] = None
+        self.info: Optional[JSON] = None
 
 
 class ErrorDetail(_serialization.Model):
@@ -285,11 +262,11 @@ class ErrorDetail(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.code = None
-        self.message = None
-        self.target = None
-        self.details = None
-        self.additional_info = None
+        self.code: Optional[str] = None
+        self.message: Optional[str] = None
+        self.target: Optional[str] = None
+        self.details: Optional[List["_models.ErrorDetail"]] = None
+        self.additional_info: Optional[List["_models.ErrorAdditionalInfo"]] = None
 
 
 class ErrorResponse(_serialization.Model):
@@ -313,7 +290,7 @@ class ErrorResponse(_serialization.Model):
         self.error = error
 
 
-class Extension(ProxyResource):  # pylint: disable=too-many-instance-attributes
+class Extension(ProxyResource):
     """The Extension object.
 
     Variables are only populated by the server, and will be ignored when sending a request.
@@ -453,7 +430,7 @@ class Extension(ProxyResource):  # pylint: disable=too-many-instance-attributes
         """
         super().__init__(**kwargs)
         self.identity = identity
-        self.system_data = None
+        self.system_data: Optional["_models.SystemData"] = None
         self.extension_type = extension_type
         self.auto_upgrade_minor_version = auto_upgrade_minor_version
         self.release_train = release_train
@@ -461,11 +438,11 @@ class Extension(ProxyResource):  # pylint: disable=too-many-instance-attributes
         self.scope = scope
         self.configuration_settings = configuration_settings
         self.configuration_protected_settings = configuration_protected_settings
-        self.provisioning_state = None
+        self.provisioning_state: Optional[Union[str, "_models.ProvisioningState"]] = None
         self.statuses = statuses
-        self.error_info = None
-        self.custom_location_settings = None
-        self.package_uri = None
+        self.error_info: Optional["_models.ErrorDetail"] = None
+        self.custom_location_settings: Optional[Dict[str, str]] = None
+        self.package_uri: Optional[str] = None
         self.aks_assigned_identity = aks_assigned_identity
 
 
@@ -499,8 +476,8 @@ class ExtensionPropertiesAksAssignedIdentity(_serialization.Model):
         :paramtype type: str
         """
         super().__init__(**kwargs)
-        self.principal_id = None
-        self.tenant_id = None
+        self.principal_id: Optional[str] = None
+        self.tenant_id: Optional[str] = None
         self.type = type
 
 
@@ -529,8 +506,8 @@ class ExtensionsList(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.value = None
-        self.next_link = None
+        self.value: Optional[List["_models.Extension"]] = None
+        self.next_link: Optional[str] = None
 
 
 class ExtensionStatus(_serialization.Model):
@@ -622,10 +599,10 @@ class ExtensionType(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.system_data = None
-        self.release_trains = None
-        self.cluster_types = None
-        self.supported_scopes = None
+        self.system_data: Optional["_models.SystemData"] = None
+        self.release_trains: Optional[List[str]] = None
+        self.cluster_types: Optional[Union[str, "_models.ClusterTypes"]] = None
+        self.supported_scopes: Optional["_models.SupportedScopes"] = None
 
 
 class ExtensionTypeList(_serialization.Model):
@@ -699,7 +676,7 @@ class ExtensionVersionList(_serialization.Model):
         super().__init__(**kwargs)
         self.versions = versions
         self.next_link = next_link
-        self.system_data = None
+        self.system_data: Optional["_models.SystemData"] = None
 
 
 class ExtensionVersionListVersionsItem(_serialization.Model):
@@ -730,7 +707,7 @@ class ExtensionVersionListVersionsItem(_serialization.Model):
         self.versions = versions
 
 
-class FluxConfiguration(ProxyResource):  # pylint: disable=too-many-instance-attributes
+class FluxConfiguration(ProxyResource):
     """The Flux Configuration object returned in Get & Put response.
 
     Variables are only populated by the server, and will be ignored when sending a request.
@@ -867,7 +844,7 @@ class FluxConfiguration(ProxyResource):  # pylint: disable=too-many-instance-att
         :paramtype configuration_protected_settings: dict[str, str]
         """
         super().__init__(**kwargs)
-        self.system_data = None
+        self.system_data: Optional["_models.SystemData"] = None
         self.scope = scope
         self.namespace = namespace
         self.source_kind = source_kind
@@ -875,13 +852,13 @@ class FluxConfiguration(ProxyResource):  # pylint: disable=too-many-instance-att
         self.git_repository = git_repository
         self.kustomizations = kustomizations
         self.configuration_protected_settings = configuration_protected_settings
-        self.statuses = None
-        self.repository_public_key = None
-        self.last_source_synced_commit_id = None
-        self.last_source_synced_at = None
-        self.compliance_state = None
-        self.provisioning_state = None
-        self.error_message = None
+        self.statuses: Optional[List["_models.ObjectStatusDefinition"]] = None
+        self.repository_public_key: Optional[str] = None
+        self.last_source_synced_commit_id: Optional[str] = None
+        self.last_source_synced_at: Optional[datetime.datetime] = None
+        self.compliance_state: Optional[Union[str, "_models.FluxComplianceState"]] = None
+        self.provisioning_state: Optional[Union[str, "_models.ProvisioningState"]] = None
+        self.error_message: Optional[str] = None
 
 
 class FluxConfigurationPatch(_serialization.Model):
@@ -975,8 +952,8 @@ class FluxConfigurationsList(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.value = None
-        self.next_link = None
+        self.value: Optional[List["_models.FluxConfiguration"]] = None
+        self.next_link: Optional[str] = None
 
 
 class GitRepositoryDefinition(_serialization.Model):
@@ -1183,8 +1160,8 @@ class Identity(_serialization.Model):
         :paramtype type: str
         """
         super().__init__(**kwargs)
-        self.principal_id = None
-        self.tenant_id = None
+        self.principal_id: Optional[str] = None
+        self.tenant_id: Optional[str] = None
         self.type = type
 
 
@@ -1468,8 +1445,8 @@ class OperationStatusList(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.value = None
-        self.next_link = None
+        self.value: Optional[List["_models.OperationStatusResult"]] = None
+        self.next_link: Optional[str] = None
 
 
 class OperationStatusResult(_serialization.Model):
@@ -1477,7 +1454,7 @@ class OperationStatusResult(_serialization.Model):
 
     Variables are only populated by the server, and will be ignored when sending a request.
 
-    All required parameters must be populated in order to send to Azure.
+    All required parameters must be populated in order to send to server.
 
     :ivar id: Fully qualified ID for the async operation.
     :vartype id: str
@@ -1528,7 +1505,7 @@ class OperationStatusResult(_serialization.Model):
         self.name = name
         self.status = status
         self.properties = properties
-        self.error = None
+        self.error: Optional["_models.ErrorDetail"] = None
 
 
 class PatchExtension(_serialization.Model):
@@ -1689,8 +1666,8 @@ class ResourceProviderOperation(_serialization.Model):
         super().__init__(**kwargs)
         self.name = name
         self.display = display
-        self.is_data_action = None
-        self.origin = None
+        self.is_data_action: Optional[bool] = None
+        self.origin: Optional[str] = None
 
 
 class ResourceProviderOperationDisplay(_serialization.Model):
@@ -1768,7 +1745,7 @@ class ResourceProviderOperationList(_serialization.Model):
         """
         super().__init__(**kwargs)
         self.value = value
-        self.next_link = None
+        self.next_link: Optional[str] = None
 
 
 class Scope(_serialization.Model):
@@ -1849,7 +1826,7 @@ class ScopeNamespace(_serialization.Model):
         self.target_namespace = target_namespace
 
 
-class SourceControlConfiguration(ProxyResource):  # pylint: disable=too-many-instance-attributes
+class SourceControlConfiguration(ProxyResource):
     """The SourceControl Configuration object returned in Get & Put response.
 
     Variables are only populated by the server, and will be ignored when sending a request.
@@ -1981,7 +1958,7 @@ class SourceControlConfiguration(ProxyResource):  # pylint: disable=too-many-ins
          ~azure.mgmt.kubernetesconfiguration.v2021_11_01_preview.models.HelmOperatorProperties
         """
         super().__init__(**kwargs)
-        self.system_data = None
+        self.system_data: Optional["_models.SystemData"] = None
         self.repository_url = repository_url
         self.operator_namespace = operator_namespace
         self.operator_instance_name = operator_instance_name
@@ -1989,12 +1966,12 @@ class SourceControlConfiguration(ProxyResource):  # pylint: disable=too-many-ins
         self.operator_params = operator_params
         self.configuration_protected_settings = configuration_protected_settings
         self.operator_scope = operator_scope
-        self.repository_public_key = None
+        self.repository_public_key: Optional[str] = None
         self.ssh_known_hosts_contents = ssh_known_hosts_contents
         self.enable_helm_operator = enable_helm_operator
         self.helm_operator_properties = helm_operator_properties
-        self.provisioning_state = None
-        self.compliance_status = None
+        self.provisioning_state: Optional[Union[str, "_models.ProvisioningStateType"]] = None
+        self.compliance_status: Optional["_models.ComplianceStatus"] = None
 
 
 class SourceControlConfigurationList(_serialization.Model):
@@ -2023,8 +2000,8 @@ class SourceControlConfigurationList(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.value = None
-        self.next_link = None
+        self.value: Optional[List["_models.SourceControlConfiguration"]] = None
+        self.next_link: Optional[str] = None
 
 
 class SupportedScopes(_serialization.Model):